setting up more of the core in CoreRuntime, this time the logger and Umbraco Core
